zhengfang_system_spider/requirements.txt
lxml==4.2.1
requests==2.18.4
Pillow==5.2.0
beautifulsoup4==4.6.0
